+++ backend/src/main/java/com/tournet/tournetERP/contents/entity/TourServiceInput.java
@@ -13,7 +13,7 @@ public class TourServiceInput {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "SERVICE_INPUT_UUID", nullable = false)
-    private int servInputUuid; //입력내용UUID
+    private long servInputUuid; //입력내용UUID
 
     @Column(name = "SERVICE_INPUT_KOR", nullable = true)
     private String servInputKor; //입력항목KOR
